Create README.md
Executive Abstract
The doctoral thesis concerns the development of a customized and user-friendly 3D model of digital documentation of the natural environment, emphasizing urban recreation parks. The proposed model is based on an automated design process that integrates multimedia data, Blockchain functionality, and "smart" statistical analysis. 
The research methodology refers to the clear definition of the requirements of digital documentation of the natural environment, the definition of the metadata of this documentation, the selection of the appropriate design platform, the development of a customized user-application interface, the development of automated and parameterized design routines for the construction of 3D geometry, and the development the relationship of the 3D geometry with Blockchain databases.
The originality of the thesis is characterized by the integration of virtual reality techniques in the development of the parameterized 3D geometry, the interactive connection of the digital documentation objects with spatial databases, and the integration of the proposed 3D model in a Blockchain environment (Internet of Trees Blockchain).

Keywords
Landscape ecology; geodesign; tree modeling architecture; Immersive virtual reality; Visuospatial cognition and orientation; Environmental geometry; Action observation; Motor Imagery; Environmental boundaries; Environmental modeling, distributed and collaborative CAD; smart forest monitoring; landscape architecture; AGB; Metaverse; Blockchain functionalities; Geometric perception; Reality modeling.
 